feat: Initialize core application structure, navigation bar, main views, friend system, and development routes. #82
